refactor(cli): extract session navigation command subsystem module (#168)
